+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Min Dart SDK 2.16.
 * **BREAKING**: `nextYear()`/`previousYear()` now saves time and month (!).
 * Method `addYears()` - Returns the `DateTime` resulting from adding the given number of years to this `DateTime`.
+* Method `addWeeks()` - Returns the `DateTime` resulting from adding the given number of weeks to this `DateTime`.
 
 ## [1.0.3] - 2022-04-22